requirements.txt
#http://downloads.sourceforge.net/project/jpype/JPype/0.5.4/JPype-0.5.4.2.zip
#-e git+git://github.com/erikrose/chardet#egg=chardet
-e git+git://github.com/misja/python-boilerpipe#egg=boilerpipe
dateutils
